implement callback.

+++ edusdk-api/src/main/kotlin/io/agora/education/api/message/EduMsg.kt
@@ -4,5 +4,5 @@ open class EduMsg(
         val userUuid: String,
         val userName: String,
         val message: String,
-        val timeStamp: Int
+        val timeStamp: Long
 )
