+++ edusdk-impl/src/main/kotlin/io/agora/rte/data/RtmResponseBody.kt
@@ -3,6 +3,6 @@ package io.agora.rte.data
 class RtmResponseBody<T>(
         val cmd: Int,
         val version: Int,
-        val ts: Int,
+        val ts: Long,
         val data: T) {
 }
